backend refactor translator, frontend move services to one folder
